# shared/contracts/openapi/iris-v1.yaml
openapi: 3.1.0
...
info:
  title: IRIS Ticketing Platform — Integration API
  version: '1.0.0'
  summary: The only external surface. Everything an integrating product can do, it does here.
  description: |
    A standalone multi-product support ticketing platform. Products integrate
    through this REST contract plus an embeddable widget; they never own ticket
    data and never build a support UI.

    **Authentication** comes in two forms:

    - **Server credential** — `client_id` + HMAC-SHA256 over a canonical request
      string. Full access to the product's own tickets.
    - **Publishable key** — safe to ship in a browser. Permits exactly four
      operations, all scoped to the bearer of the accompanying identity JWT.

    A verifiable HMAC test vector is published in `docs/api-contract.md` §3.3 so
    an integrator can self-check their signing code before their first live call.
  contact:
    name: IRIS Platform Team
  license:
    name: Proprietary
servers:
  - url: https://support.iris.example
    description: Production
  - url: http://localhost:4000
    description: Local development (the gateway)
...
security:
  - HmacSignature: []
    HmacKey: []
    HmacTimestamp: []
    HmacNonce: []
paths:
...
  /v1/widget/config:
    get:
      tags: [Widget]
      operationId: getWidgetConfig
      summary: Widget bootstrap configuration
      description: |
        Everything product-specific — visible fields, categories, severities,
        branding, enabled capabilities, suggested questions.

        **This is what makes the zero-code tier literal:** a product changes its
        widget by editing config in the admin portal, with no redeploy on its side.
      security:
        - PublishableKey: []
      parameters:
        - $ref: '#/components/parameters/Identity'
      responses:
        '200':
          description: Widget definition
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WidgetConfig' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/widget/ask:
    post:
      tags: [Widget]
      operationId: askQuestion
      summary: Ask a question (deflection)
      description: |
        Retrieval over the product's knowledge base and its **resolved** tickets.

        ⚠️ **This endpoint never creates a ticket** and never acts on one. It
        returns no ticket id. It is *deflection* — a user getting an answer and
        choosing not to file — not auto-resolution. See ADR-009.
      security:
        - PublishableKey: []
      parameters:
        - $ref: '#/components/parameters/Identity'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string, minLength: 1, maxLength: 2000 }
                conversation_id: { type: [string, 'null'], maxLength: 80 }
                context: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Candidate answers
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AskResponse' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503':
          description: |
            `deflection_unavailable` — the search backend is down. **Ticket
            creation is unaffected**; fall through to `POST /v1/tickets`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /v1/kb/articles:
    get:
      tags: [Knowledge base]
      operationId: searchKbArticles
      summary: Search the knowledge base
      description: Only `published` articles in the product's public categories are ever returned.
      security:
        - PublishableKey: []
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Full-text + trigram hybrid }
        - { name: category, in: query, schema: { type: string } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Identity'
      responses:
        '200':
          description: Matching articles
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbArticlePage' }
        '403':
          description: '`knowledge_base_disabled` — not enabled for this product'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /v1/kb/articles/{id}:
    get:
      tags: [Knowledge base]
      operationId: getKbArticle
      summary: Retrieve one article
      security:
        - PublishableKey: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/Identity'
      responses:
        '200':
          description: The article, including its body
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbArticle' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/kb/articles/{id}/helpful:
    post:
      tags: [Knowledge base]
      operationId: voteArticleHelpful
      summary: Record a helpful vote
      security:
        - PublishableKey: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/Identity'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [helpful]
              properties:
                helpful: { type: boolean }
      responses:
        '200':
          description: Vote recorded
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/announcements:
    get:
      tags: [Widget]
      operationId: listAnnouncements
      summary: Product announcements
      security:
        - PublishableKey: []
      responses:
        '200':
          description: Active announcements, newest first
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Announcement' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    HmacKey:
      type: apiKey
      in: header
      name: X-IRIS-Key
      description: Your `client_id`.
    HmacTimestamp:
      type: apiKey
      in: header
      name: X-IRIS-Timestamp
      description: Unix seconds. Must be within ±300s of platform time.
    HmacNonce:
      type: apiKey
      in: header
      name: X-IRIS-Nonce
      description: UUIDv4, unique per request. Cached 600s.
    HmacSignature:
      type: apiKey
      in: header
      name: X-IRIS-Signature
      description: |
        `v1=<hex>` where hex is HMAC-SHA256 over the canonical string:

            v1 \n METHOD \n PATH_WITH_QUERY \n TIMESTAMP \n NONCE \n sha256_hex(RAW_BODY)

        Method and path are inside the signature, so a captured request cannot
        be replayed against a different endpoint.
    PublishableKey:
      type: apiKey
      in: header
      name: X-IRIS-Publishable-Key
      description: |
        `pub_live_…`. Public by design and therefore low-privilege: it permits
        only ticket creation, reading the bearer's own tickets, asking a
        question, and searching the knowledge base — enforced against the
        product's origin allowlist and tightly rate limited.
  parameters:
...
    Identity:
      name: X-IRIS-Identity
      in: header
      required: false
      description: |
        A short-lived JWT the integrating product minted for the end user
        (RS256/ES256, `exp − iat ≤ 300s`, audience `iris-ticketing`). Required
        when acting on behalf of a raiser.
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
...
    Unauthorized:
      description: '`signature_invalid`, `timestamp_out_of_window`, `nonce_replayed` or `identity_token_invalid`'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: '`credential_scope_exceeded` or `origin_not_allowed`'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: |
        `ticket_not_found`. Deliberately indistinguishable from "not yours" — a
        403 would confirm the resource exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    RateLimited:
      description: '`rate_limited`. Honour `Retry-After`.'
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    Severity:
      type: string
      enum: [low, medium, high, critical]
      description: Stored and published as `severity`. Some UIs label the column "Priority"; that is display only.
...
    ErrorEnvelope:
      type: object
      required: [error]
      description: |
        Every non-2xx response, without exception, has this shape. Branch on
        `code` — never on `message`, which may be reworded without a version bump.
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unsupported_category
                - signature_invalid
                - timestamp_out_of_window
                - nonce_replayed
                - identity_token_invalid
                - identity_provider_unreachable
                - unauthenticated
                - credential_scope_exceeded
                - origin_not_allowed
                - knowledge_base_disabled
                - forbidden
                - ticket_not_found
                - not_found
                - invalid_state_transition
                - idempotency_key_reuse
                - already_exists
                - attachment_too_large
                - attachment_type_not_allowed
                - preauth_grant_invalid
                - rate_limited
                - internal_error
                - service_unavailable
                - deflection_unavailable
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
...
    AskAnswer:
      type: object
      required: [type, id, title, excerpt, score]
      properties:
        type: { type: string, enum: [kb_article, resolved_ticket] }
        id: { type: string }
        title: { type: string }
        excerpt: { type: string }
        score: { type: number }
        url: { type: string }
    AskResponse:
      type: object
      required: [conversation_id, suggested_action, answers, prefill]
      properties:
        conversation_id: { type: string, examples: ['cnv_01JQZ9A8S6'] }
        suggested_action:
          type: string
          enum: [answer, create_ticket]
          description: |
            `create_ticket` when the best score is below the product's
            `deflection.min_score`. A weak answer is worse than no answer.
        answers:
          type: array
          items: { $ref: '#/components/schemas/AskAnswer' }
        prefill:
          type: object
          required: [description]
          description: Carried into `POST /v1/tickets` if the user proceeds.
          properties:
            description: { type: string }
            category: { type: [string, 'null'] }
            severity:
              oneOf:
                - $ref: '#/components/schemas/Severity'
                - type: 'null'
    KbArticle:
      type: object
      required: [id, title, excerpt, helpful_pct, views]
      properties:
        id: { type: string }
        title: { type: string }
        category: { type: [string, 'null'] }
        excerpt: { type: string }
        body: { type: string, description: Present only on single-article retrieval }
        score: { type: number, description: Present only on search }
        helpful_pct: { type: [integer, 'null'], minimum: 0, maximum: 100 }
        views: { type: integer, minimum: 0 }
    KbArticlePage:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/KbArticle' }
        next_cursor: { type: [string, 'null'] }
        has_more: { type: boolean }
    Announcement:
      type: object
      required: [id, title, body, kind, published_at]
      properties:
        id: { type: string }
        title: { type: string }
        body: { type: string }
        kind: { type: string, enum: [info, maintenance, incident, release] }
        published_at: { type: string, format: date-time }
    WidgetConfig:
      type: object
      required: [product, branding, capabilities, fields, categories, severities, disclaimer, footer]
      description: |
        Everything the widget renders. A product changes its widget by editing
        this server-side, with no redeploy on its side.
      properties:
        product:
          type: object
          required: [id, name, slug]
          properties:
            id: { type: string }
            name: { type: string }
            slug: { type: string }
        branding:
          type: object
          required: [title, subtitle, greeting, primary_color, accent_color]
          properties:
            title: { type: string }
            subtitle: { type: string }
            greeting: { type: string }
            primary_color: { type: string }
            accent_color: { type: string }
            logo_text: { type: [string, 'null'] }
        capabilities:
          type: array
          description: Which of the eight tiles appear.
          items:
            type: string
            enum: [ask, create_ticket, search_docs, my_tickets, upload_screenshot, live_chat, ai_suggestions, announcements]
        fields:
          type: object
          properties:
            subject: { type: boolean }
            category: { type: boolean }
            severity: { type: boolean }
            attachments: { type: boolean }
        categories:
          type: array
          items:
            type: object
            required: [value, label]
            properties:
              value: { type: string }
              label: { type: string }
        severities:
          type: array
          items:
            type: object
            required: [value, label]
            properties:
              value: { type: string }
              label: { type: string }
        default_severity: { type: string }
        suggestions: { type: array, items: { type: string } }
        knowledge_base_enabled: { type: boolean }
        deflection_enabled: { type: boolean }
        allow_anonymous: { type: boolean }
        identity:
          type: object
          required: [authenticated]
          properties:
            authenticated: { type: boolean }
            name: { type: [string, 'null'] }
            email: { type: [string, 'null'] }
        disclaimer: { type: string }
        footer: { type: string }